Move url update from Search to Results
